Add caching protocol for pocket and contiles
